requirements.txt
customtkinter
tkinter
pillow